# stellar-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar API
  description: |
    # Stellar API Documentation
    
    Modern REST API with JWT authentication, role-based access control, and comprehensive security features.
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Rate limiting and security headers
    - Input validation and sanitization
    - Comprehensive error handling
    - API versioning
    - Two-factor authentication support
    - Account lockout protection
    
    ## Authentication
    
    This API uses JWT (JSON Web Tokens) for authentication. To access protected endpoints:
    
    1. Register a new account or login with existing credentials
    2. Obtain an access token from the `/auth/login` endpoint
    3. Include the token in the `Authorization` header: `Bearer <token>`
    4. Use refresh tokens to obtain new access tokens when they expire
    
    ## Rate Limiting
    
    API requests are rate-limited to prevent abuse:
    - Default: 100 requests per 15 minutes per IP
    - Authentication endpoints: 5 requests per 15 minutes
    - Password reset: 3 requests per hour
    
    ## Error Responses
    
    All errors follow a consistent format:
    ```json
    {
      "error": "Error type",
      "message": "Human-readable error message",
      "statusCode": 400,
      "details": {} // Optional additional information
    }
    ```
    
  version: 1.0.0
  contact:
    name: Stellar Team
...
  license:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          format: objectId
          example: "507f1f77bcf86cd799439011"
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        profile:
          $ref: '#/components/schemas/UserProfile'
        emailVerified:
          type: boolean
          example: true
        status:
          type: string
          enum: [active, inactive, suspended, pending]
          example: "active"
        role:
          type: string
          enum: [user, admin, moderator]
          example: "user"
        lastLogin:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 50
          example: "John"
        lastName:
          type: string
          maxLength: 50
          example: "Doe"
        fullName:
          type: string
          readOnly: true
          example: "John Doe"
        avatar:
          type: string
          format: uri
          nullable: true
        bio:
          type: string
          maxLength: 500
          example: "Software developer passionate about building great APIs"
        dateOfBirth:
          type: string
          format: date
        phoneNumber:
          type: string
          pattern: "^\\+?[\\d\\s-()]+$"
          example: "+1-555-123-4567"
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email or username
          example: "john@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "SecurePassword123!"
        rememberMe:
          type: boolean
          default: false
          description: Generate longer-lived refresh token
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        tokens:
          type: object
          properties:
            accessToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            refreshToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            expiresIn:
              type: integer
              description: Access token expiry time in seconds
              example: 3600
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          minLength: 8
          description: Must contain uppercase, lowercase, number, and special character
          example: "SecurePassword123!"
        profile:
          type: object
          properties:
            firstName:
              type: string
              example: "John"
            lastName:
              type: string
              example: "Doe"
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          example: "CurrentPassword123!"
        newPassword:
          type: string
          format: password
          minLength: 8
          example: "NewSecurePassword123!"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid request parameters"
        statusCode:
          type: integer
          example: 400
        details:
          type: object
          additionalProperties: true
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email verification
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User registered successfully. Please check your email to verify your account."
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive access and refresh tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate current refresh token and logout user
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Optional refresh token to invalidate
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Logged out successfully"
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Use refresh token to obtain new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tokens:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: integer
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/change-password:
    post:
      tags:
        - Users
      summary: Change password
      description: Change the authenticated user's password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password changed successfully"
        '401':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a user's public profile by their ID (admin only)
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: objectId
          description: User ID
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: Get a paginated list of users (admin only)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, -createdAt, username, -username, email, -email]
            default: -createdAt
          description: Sort field and order
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended, pending]
          description: Filter by status
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin, moderator]
          description: Filter by role
        - name: search
          in: query
          schema:
            type: string
          description: Search in username and email
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 100
                      pages:
                        type: integer
                        example: 5